moved ark custom components to arkportal.  added images to home page, and fixed favicon
