--- src/portal-components/ark/ARKWelcomePage.tsx
@@ -1,28 +1,0 @@
-import * as React from 'react'
-import { Typography } from 'synapse-react-client'
-
-const ARKWelcomePage = () => {
-  return (
-    <div className="ARKWelcomePage container-fluid" style={{ paddingTop: 50, paddingBottom: 50 }}>
-      <div className="row">
-        <div className="col-md-offset-1 col-md-10">
-          <p style={{ fontSize: 50, fontWeight: 700 }}>
-            Welcome to the ARK Portal
-          </p>
-          <div className="description">
-            <Typography variant="body1">
-              The ARK Portal hosts data generated by a network of research teams working collaboratively to deepen the understanding of Autoimmune and Immune-Mediated Diseases
-            </Typography>
-          </div>
-          <div className="funderNote">
-            <Typography variant="body1">
-              Established by the Accelerating Medicines Partnership® (AMP®)
-            </Typography>
-          </div>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-export default ARKWelcomePage
