+++ src/portal-components/arkportal/ARKWelcomePage.tsx
@@ -1,0 +1,35 @@
+import Layout from 'portal-components/Layout'
+import * as React from 'react'
+import { Typography } from 'synapse-react-client'
+import logoPath from './assets/arklogomark.png'
+
+const ARKWelcomePage = () => {
+  return (
+    <>
+      <Layout containerClassName='ARKWelcomePage'>
+          <div className="flexLayout"> 
+            <div className="mainWelcomeContent">
+              <p style={{ fontSize: 50, fontWeight: 700 }}>
+                Welcome to the ARK Portal
+              </p>
+              <div className="description">
+                <Typography variant="body1">
+                  The ARK Portal hosts data generated by a network of research teams working collaboratively to deepen the understanding of Autoimmune and Immune-Mediated Diseases
+                </Typography>
+              </div>
+              <div className="funderNote">
+                <Typography variant="body1">
+                  Established by the Accelerating Medicines Partnership® (AMP®)
+                </Typography>
+              </div>
+            </div>
+            <img className="welcomeARKLogo" src={logoPath} alt="ARK Portal logo"/>
+          </div>
+      </Layout>
+      <img className="lowerRightARKLogo" src={logoPath} alt="ARK Portal logo"/>
+      <img className="lowerLeftARKLogo" src={logoPath} alt="ARK Portal logo"/>
+    </>
+  )
+}
+
+export default ARKWelcomePage
