+++ src/portal-components/arkportal/style/_ARKWelcomePage.scss
@@ -1,0 +1,41 @@
+@use 'variables' as Portal;
+
+.ARKWelcomePage {
+  padding-top: 50px;
+  padding-bottom: 70px;
+  .flexLayout {
+    display: flex;
+    justify-content: space-between;
+    flex-wrap: wrap;
+  }
+  .mainWelcomeContent {
+    max-width: 700px;
+  }
+  .welcomeARKLogo {
+    max-width: 200px;
+    max-height: 200px;
+    opacity: 0.25;
+    transform: rotate(10deg);
+    margin-top: 20px;
+    margin-left: 50px;
+  }
+}
+.lowerRightARKLogo {
+  position: absolute;
+  right: -200px;
+  opacity: 0.25;
+  z-index: -1;
+  max-width: 500px;
+  bottom: -100px;
+  transform: rotate(-110deg);
+}
+
+.lowerLeftARKLogo {
+  position: absolute;
+  left: -100px;
+  opacity: 0.25;
+  z-index: -1;
+  max-width: 200px;
+  bottom: -50px;
+  transform: matrix(-0.34, 0.94, 0.94, 0.34, 0, 0);
+}
