+++ src/portal-components/index.ts
@@ -7,7 +7,7 @@ import Ecosystem from './csbc-home-page/Ecosystem'
 import AboutPortal from './csbc-home-page/AboutPortal'
 import DevelopedBySage from './csbc-home-page/DevelopedBySage'
 import BrowseToolsPage from './nf/BrowseToolsPage'
-import ARKWelcomePage from './ark/ARKWelcomePage'
+import ARKWelcomePage from './arkportal/ARKWelcomePage'
 import ParticipantsBarPlot from './crc-researcher/ParticipantsBarPlot'
 import StatusLineChart from './crc-researcher/StatusLineChart'
 import SurveysCompletedPlots from './crc-researcher/SurveysCompletedPlots'
